# api/openapi.yaml
openapi: 3.1.0
info:
  title: hostd Phase A API
  version: 0.1.0
  description: Authenticated local-only control plane. Workload execution is disabled by default and requires an explicit runtime capability.
servers:
  - url: http://127.0.0.1:7345
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/apps/{appId}/configuration:
    parameters: [{$ref: '#/components/parameters/AppId'}]
    get:
      operationId: getApplicationConfiguration
      responses:
        '200': {description: Current application configuration metadata and visible variables. Secret values are omitted., headers: {Cache-Control: {schema: {type: string, const: no-store}}}, content: {application/json: {schema: {$ref: '#/components/schemas/ApplicationConfiguration'}}}}
        default: {$ref: '#/components/responses/Problem'}
    put:
      operationId: replaceApplicationConfiguration
      security: [{sessionCookie: [], csrfHeader: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ReplaceApplicationConfigurationRequest'}}}}
      responses:
        '200': {description: New immutable configuration revision, headers: {Cache-Control: {schema: {type: string, const: no-store}}}, content: {application/json: {schema: {$ref: '#/components/schemas/ApplicationConfiguration'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hostd_session
      description: Opaque random session token. Set HttpOnly, Path=/, SameSite=Strict; Secure is enabled when served over TLS. Only a SHA-256 hash is stored.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with sessionCookie for every non-safe request. Rotated by the authenticated CSRF endpoint; only its hash is stored.
...
  parameters:
    AppId: {name: appId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    Problem:
      description: RFC 9457 problem response
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, code, requestId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code: {type: string}
        requestId: {type: string}
        errors: {type: object, additionalProperties: {type: string}}
...
    ConfigurationEntry: {type: object, required: [key, sensitive], properties: {key: {type: string}, sensitive: {type: boolean}, value: {type: string}, phase: {type: string, enum: [runtime, build, migration]}, targetComponent: {type: string}}}
    ConfigurationValueInput: {type: object, additionalProperties: false, required: [key, value], properties: {key: {type: string, pattern: '^[A-Za-z_][A-Za-z0-9_]*$', maxLength: 128}, value: {type: string, maxLength: 8192}}}
    SecretConfigurationValueInput: {type: object, additionalProperties: false, required: [key, value], properties: {key: {type: string, pattern: '^[A-Za-z_][A-Za-z0-9_]*$', maxLength: 128}, value: {type: string, minLength: 1, maxLength: 8192}}}
    ApplicationConfiguration: {type: object, required: [revisionNumber, entries], properties: {revisionId: {type: string, format: uuid}, revisionNumber: {type: integer, format: int64, minimum: 0}, formatVersion: {type: integer, enum: [1, 2]}, deploymentPlanRevisionId: {type: string, format: uuid}, deploymentPlanRevisionNumber: {type: integer, format: int64, minimum: 0}, updatedAt: {type: string, format: date-time}, entries: {type: array, items: {$ref: '#/components/schemas/ConfigurationEntry'}}}}
    ReplaceApplicationConfigurationRequest: {type: object, additionalProperties: false, required: [expectedRevisionNumber, variables, secrets, remove], properties: {expectedRevisionNumber: {type: integer, format: int64, minimum: 0}, variables: {type: array, maxItems: 256, items: {$ref: '#/components/schemas/ConfigurationValueInput'}}, secrets: {type: array, maxItems: 256, items: {$ref: '#/components/schemas/SecretConfigurationValueInput'}}, remove: {type: array, maxItems: 256, items: {type: string, pattern: '^[A-Za-z_][A-Za-z0-9_]*$', maxLength: 128}}}}
...